+++ src/fable-library-rust/src/System.Threading.fs
@@ -1,0 +1,5 @@
+namespace System.Threading
+
+open Global_
+
+type CancellationToken = struct end