+++ src/Fable.Cli/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+* [Rust] Added basic class inheritance support (by @ncave)
+* [Rust] Added String.Replace(char, char) and test (by @ncave)
+
 ## 4.22.0 - 2024-10-02
 
 ### Added